+++ src/utils.c
@@ -1,6 +1,6 @@
 #include "include/utils.h"
 #include <ctype.h>
-#include <stdint.h>
+#include <inttypes.h>
 #include <stdio.h>
 #include <string.h>
 #include <sys/time.h>
@@ -10,23 +10,23 @@ static const uint8_t tab64[64] = {63, 0,  58, 1,  59, 47, 53, 2,  60, 39, 48, 27
                                   62, 57, 46, 52, 38, 26, 32, 41, 50, 36, 17, 19, 29, 10, 13, 21,
                                   56, 45, 25, 31, 35, 16, 9,  12, 44, 24, 15, 8,  23, 7,  6,  5};
 
-static uint8_t log2_64(uint64_t value) {
+static uint8_t log2_64(unsigned long long value) {
     value |= value >> 1;
     value |= value >> 2;
     value |= value >> 4;
     value |= value >> 8;
     value |= value >> 16;
     value |= value >> 32;
-    return tab64[((uint64_t) ((value - (value >> 1)) * 0x07EDD5E59A4E28C2)) >> 58];
+    return tab64[((unsigned long long) ((value - (value >> 1)) * 0x07EDD5E59A4E28C2)) >> 58];
 }
 
-uint8_t utils_bit_bitscan_forward(uint64_t* x) {
+uint8_t utils_bit_bitscan_forward(unsigned long long* x) {
     uint8_t position;
 #ifdef __GNUC__
     position = __builtin_ctzll(*x);
 #else
-    uint64_t ls1b = *x & -*x;
-    position      = log2_64(ls1b);
+    unsigned long long ls1b = *x & -*x;
+    position                = log2_64(ls1b);
 #endif
     *x &= *x - 1;
     return position;
@@ -42,7 +42,7 @@ uint8_t utils_bit_count1s(int64_t n) {
     return count;
 }
 
-uint64_t utils_bit_pop(const int64_t n, const uint8_t k) {
+unsigned long long utils_bit_pop(const int64_t n, const uint8_t k) {
     int64_t b = 1ULL << k;
 
     if (n & b)
@@ -87,10 +87,10 @@ char* utils_str_trimwhitespace(char* str) {
     return str;
 }
 
-uint64_t utils_time_curr_time_ms(void) {
+unsigned long long utils_time_curr_time_ms(void) {
     struct timeval t;
     gettimeofday(&t, NULL);
-    return (uint64_t) (t.tv_sec * 1000) + t.tv_usec / 1000;
+    return (unsigned long long) (t.tv_sec * 1000) + t.tv_usec / 1000;
 }
 
 void utils_bin_display32(const int32_t x) {
